oauth/pyproject.toml
[project]
name = "oauth"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
oauth = "oauth.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oauth"]

[tool.hatch.build.targets.wheel.force-include]
"src/oauth/providers.yaml" = "oauth/providers.yaml"
